% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{anchorOverlap}
\alias{anchorOverlap}
\title{Identifies Overlapping Anchors}
\usage{
anchorOverlap(rep1.anchor, rep2.anchor, max.gap = 1000L)
}
\arguments{
\item{rep1.anchor}{data frame with the following columns:
\tabular{rl}{
  \code{chr} \tab character; genomic location of anchor in replicate 1 -
  chromosome (e.g., \code{"chr3"})\cr
  \code{start} \tab integer; genomic location of anchor in replicate 1 -
  start coordinate\cr
  \code{end} \tab integer; genomic location of anchor in replicate 1 -
  end coordinate
}}

\item{rep2.anchor}{data frame with the following columns:
\tabular{rl}{
  \code{chr} \tab character; genomic location of anchor in replicate 2 -
  chromosome (e.g., \code{"chr3"})\cr
  \code{start} \tab integer; genomic location of anchor in replicate 2 -
  start coordinate\cr
  \code{end} \tab integer; genomic location of anchor in replicate 2 -
  end coordinate
}}

\item{max.gap}{integer; maximum gap in nucleotides allowed between two anchors for
them to be considered as overlapping (defaults to \code{1000L})}
}
\value{
A data frame with the following columns:
\tabular{rl}{
  \code{"queryHits"} \tab TODO\cr
  \code{"subjectHits"} \tab TODO
}
}
\description{
TODO
}
\examples{
TODO

}
