% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{calculateMidpointDistance}
\alias{calculateMidpointDistance}
\title{Distance between Anchor Midpoints of two Interactions}
\usage{
calculateMidpointDistance(int1.anchor.a.start, int1.anchor.a.end,
  int1.anchor.b.start, int1.anchor.b.end, int2.anchor.a.start,
  int2.anchor.a.end, int2.anchor.b.start, int2.anchor.b.end)
}
\arguments{
\item{int1.anchor.a.start}{integer vector; genomic start coordinate(s)
of anchor A in replicate 1 interaction}

\item{int1.anchor.a.end}{integer vector; genomic end coordinate(s)
of anchor A in replicate 1 interaction}

\item{int1.anchor.b.start}{integer vector; genomic start coordinate(s)
of anchor B in replicate 1 interaction}

\item{int1.anchor.b.end}{integer vector; genomic end coordinate(s)
of anchor B in replicate 1 interaction}

\item{int2.anchor.a.start}{integer vector; genomic start coordinate(s)
of anchor A in replicate 2 interaction}

\item{int2.anchor.a.end}{integer vector; genomic end coordinate(s)
of anchor A in replicate 2 interaction}

\item{int2.anchor.b.start}{integer vector; genomic start coordinate(s)
of anchor B in replicate 2 interaction}

\item{int2.anchor.b.end}{integer vector; genomic end coordinate(s)
of anchor B in replicate 2 interaction}
}
\value{
positive integer vector; distances between interaction pairs
}
\description{
Calculates the distance in nucleotides between the anchor midpoints of
two interactions, which is the sum of the distance between midpoints of
anchor A in interaction 1 and anchor A in interaction 2, and the distance
between midpoints of
anchor B in interaction 1 and anchor B in interaction 2.
}
\examples{
# identical, zero distance
calculateMidpointDistance(100, 120, 240, 260,
                          100, 120, 240, 260)

# centered, zero distance
calculateMidpointDistance(100, 120, 240, 260,
                          90, 130, 230, 270)

# off by 10 per anchor
calculateMidpointDistance(100, 120, 240, 250,
                         110, 130, 230, 240)

# off by 10 (anchor B only)
calculateMidpointDistance(100, 120, 240, 250,
                         90, 130, 250, 260)

# vectorized example
calculateMidpointDistance(c(100, 100, 100, 100),
                          c(120, 120, 120, 120),
                          c(240, 240, 240, 240),
                          c(260, 260, 250, 250),
                          c(100, 90, 110, 90),
                          c(120, 130, 130, 130),
                          c(240, 230, 230, 250),
                          c(260, 270, 240, 260))

}
