% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{calculateMidpointDistance1D}
\alias{calculateMidpointDistance1D}
\title{Distance between Midpoints of two Peaks}
\usage{
calculateMidpointDistance1D(peak1.start, peak1.end, peak2.start, peak2.end)
}
\arguments{
\item{peak1.start}{integer vector; genomic start coordinate(s)
of peak in replicate 1}

\item{peak1.end}{integer vector; genomic end coordinate(s)
of peak in replicate 1}

\item{peak2.start}{integer vector; genomic start coordinate(s)
of peak in replicate 2}

\item{peak2.end}{integer vector; genomic end coordinate(s)
of peak in replicate 2}
}
\value{
positive integer vector; distances between peak pairs
}
\description{
Calculates the distance in nucleotides between the midpoints of
two peaks.

Note: peaks must be on the same
chromosome; start coordinate is always less than end coordinate
}
\examples{
# identical, zero distance
calculateMidpointDistance1D(100, 120,
                          100, 120)

# centered, zero distance
calculateMidpointDistance1D(100, 120,
                          90, 130)

# off by 10 per anchor
calculateMidpointDistance1D(100, 120,
                         110, 130)

# vectorized example
calculateMidpointDistance1D(c(100, 100, 100),
                          c(120, 120, 120),
                          c(100, 90, 110),
                          c(120, 130, 130))

}
