% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{calculateRelativeOverlap1D}
\alias{calculateRelativeOverlap1D}
\title{Relative Anchor Overlap of two Peaks}
\usage{
calculateRelativeOverlap1D(peak1.start, peak1.end, peak2.start, peak2.end)
}
\arguments{
\item{peak1.start}{integer vector; genomic start coordinate(s)
of peak in replicate 1}

\item{peak1.end}{integer vector; genomic end coordinate(s)
of peak in replicate 1}

\item{peak2.start}{integer vector; genomic start coordinate(s)
of peak in replicate 2}

\item{peak2.end}{integer vector; genomic end coordinate(s)
of peak in replicate 2}
}
\value{
numeric vector; relative overlaps between peak pairs
}
\description{
Calculates the overlap between anchor A of interaction 1 and anchor
A of interaction 2, as well as anchor B of interaction 1 and anchor B of
interaction 2. The overlap (in nucleotides) is then normalized by the length
of the anchors.
}
\examples{
# 100\% overlap
calculateRelativeOverlap1D(100, 120,
                         100, 120)

# 50\% overlap
calculateRelativeOverlap1D(100, 120,
                         100, 110)

# negative overlap
calculateRelativeOverlap1D(100, 120,
                         130, 140)

# larger negative overlap
calculateRelativeOverlap1D(100, 120,
                         200, 220)

# vectorized example
calculateRelativeOverlap1D(c(100, 100, 100, 100),
                         c(120, 120, 120, 120),
                         c(100, 100, 130, 200),
                         c(120, 110, 140, 220))
}
