% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chiapet}
\alias{chiapet}
\title{Example Genomic Interaction Data Set}
\format{A list with two components, the data frames \code{rep1.df} and
\code{rep2.df}, which have the following seven columns:
\tabular{rl}{
  \code{chrA} \tab genomic location of anchor A - chromosome (e.g.,
  \code{"chr3"})\cr
  \code{startA} \tab genomic location of anchor A -
  start coordinate \cr
  \code{endA} \tab genomic location of anchor A -
  end coordinate\cr
  \code{chrB} \tab genomic location of anchor B -
  chromosome (e.g., \code{"chr3"})\cr
  \code{startB} \tab genomic location of anchor B -
  start coordinate\cr
  \code{endB} \tab genomic location of anchor B -
  end coordinate\cr
  \code{fdr} False Discovery Rate - significance of interaction\tab
}}
\usage{
chiapet
}
\description{
This object contains genomic interactions on chromosomes 1 to 5, which could
be the results of HiC or ChIA-PET experiments, done in duplicates.
}
\keyword{datasets}
