% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{draw_value_idr_scatterplot}
\alias{draw_value_idr_scatterplot}
\title{Create scatterplot of IDR values}
\usage{
draw_value_idr_scatterplot(df, remove_na = TRUE,
  remove_outliers = TRUE, xlab = "transformed value in replicate 1",
  ylab = "transformed value in replicate 2", log_axes = FALSE,
  log_idr = FALSE, title = "value - IDR dependence",
  color_gradient = c("rainbow", "default"), max_points_shown = 2500)
}
\arguments{
\item{df}{part of output of \code{\link{estimate_idr}}, a data frame with at
least the following named columns:
\tabular{rl}{
  \code{value} \tab numeric; p-value, FDR, or heuristic used
  to rank the peaks\cr
  \code{rep_value} \tab numeric; value of corresponding
  replicate peak\cr
  \code{idr} \tab IDR of the peak and the
  corresponding peak in the other replicate.
}}

\item{remove_na}{logical; should NA values be removed?}

\item{remove_outliers}{logical; removes extreme data points}

\item{xlab}{character; x axis label}

\item{ylab}{character; y axis label}

\item{log_axes}{logical; show logarithmized values from replicate 1 and 2
(default value is FALSE)}

\item{log_idr}{logical; use logarithmized IDRs for colors to better
distinguish highly significant IDRs
(default value is FALSE)}

\item{title}{character; plot title}

\item{color_gradient}{character; either "rainbow" or "default"}

\item{max_points_shown}{integer; default is 2500}
}
\value{
ggplot2 object; IDR value scatterplot
}
\description{
Creates diagnostic plots to visualize the results of
\code{\link{estimate_idr}}.
}
\examples{
idr_results <- estimate_idr1d(idr2d:::chipseq$rep1_df,
                              idr2d:::chipseq$rep2_df,
                              value_transformation = "log")
draw_value_idr_scatterplot(idr_results$rep1_df)

}
