% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{establishBijection2D}
\alias{establishBijection2D}
\title{Finds One-to-One Correspondence between Interactions from Replicate 1
and 2}
\usage{
establishBijection2D(rep1.df, rep2.df,
  ambiguity.resolution.method = c("value", "overlap", "midpoint"),
  max.gap = 1000L)
}
\arguments{
\item{rep1.df}{data frame of observations (i.e., genomic interactions) of
replicate 1, with at least the following columns (position of columns
matter, column names are irrelevant):
\tabular{rl}{
  column 1 (\code{chrA}) \tab character; genomic location of anchor A -
  chromosome (e.g., \code{"chr3"})\cr
  column 2 \code{startA}) \tab integer; genomic location of anchor A -
  start coordinate\cr
  column 3 (\code{endA}) \tab integer; genomic location of anchor A -
  end coordinate\cr
  column 4 (\code{chrB}) \tab character; genomic location of anchor B -
  chromosome (e.g., \code{"chr3"})\cr
  column 5 (\code{startB}) \tab integer; genomic location of anchor B -
  start coordinate\cr
  column 6 (\code{endB}) \tab integer; genomic location of anchor B -
  end coordinate\cr
  column 7 (\code{value}) \tab numeric; p-value, FDR, or heuristic used to
  rank the interactions
}}

\item{rep2.df}{data frame of observations (i.e., genomic interactions) of
replicate 2, with the following columns (position of columns
matter, column names are irrelevant):
\tabular{rl}{
  column 1 (\code{chrA}) \tab character; genomic location of anchor A -
  chromosome (e.g., \code{"chr3"})\cr
  column 2 \code{startA}) \tab integer; genomic location of anchor A -
  start coordinate\cr
  column 3 (\code{endA}) \tab integer; genomic location of anchor A -
  end coordinate\cr
  column 4 (\code{chrB}) \tab character; genomic location of anchor B -
  chromosome (e.g., \code{"chr3"})\cr
  column 5 (\code{startB}) \tab integer; genomic location of anchor B -
  start coordinate\cr
  column 6 (\code{endB}) \tab integer; genomic location of anchor B -
  end coordinate\cr
  column 7 (\code{value}) \tab numeric; p-value, FDR, or heuristic used to
  rank the interactions
}}

\item{ambiguity.resolution.method}{defines how ambiguous assignments
(when one interaction in replicate 1 overlaps with multiple interactions in
replicate 2 or vice versa)
are resolved. Available methods:
\tabular{rl}{
  \code{"value"} \tab interactions are prioritized by ascending or descending
  \code{value} column (see \code{sorting.direction}), e.g., if two
  interactions in replicate 1 overlap with one interaction in replicate 2,
  the interaction from replicate 1 is chosen which has a lower (if
  \code{sorting.direction} is \code{"ascending"}) or higher (if
  \code{"descending"}) value \cr
  \code{"overlap"} \tab the interaction pair is chosen which has the highest
  relative overlap, i.e., overlap in nucleotides of replicate 1 interaction
  anchor A and replicate 2 interaction anchor A,
  plus replicate 1 interaction anchor B and replicate 2 interaction anchor B,
  normalized by their lengths\cr
  \code{"midpoint"} \tab the interaction pair is chosen which has the
  smallest
  distance between their anchor midpoints, i.e., distance from midpoint of
  replicate 1 interaction anchor A to midpoint of
  replicate 2 interaction anchor A, plus distance from midpoint of
  replicate 1 interaction anchor B to midpoint of
  replicate 2 interaction anchor B
}}

\item{max.gap}{integer; maximum gap in nucleotides allowed between two
anchors for
them to be considered as overlapping (defaults to \code{1000L})}
}
\value{
Original data frames \code{rep1.df} and \code{rep2.df} with
additional column \code{replicate.idx}, which specifies the index of the
corresponding interaction in the other replicate. If no corresponding
interaction was found, \code{replicate.idx} is set to \code{NA}.
}
\description{
This method establishes a bijective assignment between interactions from
replicate 1 and 2. An interaction in replicate 1 is assigned to an
interaction in replicate 2 if and only if (1) both anchors of the
interactions
overlap (or the gap between anchor A/B in replicate 1 and 2 is less than
or equal to \code{max.gap}), and (2) there is no other interaction in
replicate 2 that overlaps with the interaction in replicate 1 and has a
lower \emph{ambiguity resolution value}.
}
\examples{
rep1.df <- idr2d:::chiapet$rep1.df
rep1.df$fdr <- preprocess(rep1.df$fdr, "log.additive.inverse")

rep2.df <- idr2d:::chiapet$rep2.df
rep2.df$fdr <- preprocess(rep2.df$fdr, "log.additive.inverse")

mapping <- establishBijection2D(rep1.df, rep2.df)

}
