% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hic.R
\name{estimate_idr2d_hic}
\alias{estimate_idr2d_hic}
\title{Estimates IDR for Genomic Interactions measured by HiC experiments}
\usage{
estimate_idr2d_hic(rep1_hic_file, rep2_hic_file, resolution = 1e+06,
  normalization = c("NONE", "VC", "VC_SQRT", "KR"), chromosomes = NULL,
  combined_min_value = 30, combined_max_value = Inf,
  max_factor = 1.5, jitter_factor = 1e-04, mu = 0.1, sigma = 1,
  rho = 0.2, p = 0.5, eps = 0.001, max_iteration = 30,
  local_idr = TRUE, use_python = NULL, use_virtualenv = NULL,
  use_condaenv = NULL)
}
\arguments{
\item{rep1_hic_file}{path to .hic file for replicate 1
(either local file path or URL)}

\item{rep2_hic_file}{path to .hic file for replicate 2
(either local file path or URL)}

\item{resolution}{block resolution of HiC contact matrix in base pairs,
defaults to 1,000,000 bp (usually one of the following:
2500000, 1000000, 500000, 250000, 100000, 50000, 25000, 10000, 5000)}

\item{normalization}{normalization step performed by Python package
\code{hic-straw}, one of the following: \code{"NONE"}, \code{"VC"},
\code{"VC_SQRT"}, \code{"KR"}.}

\item{chromosomes}{list of chromosome names in HiC files, defaults to UCSC
human chromosome names (chr1, ..., chr22, chrX, chrY, chrM)}

\item{combined_min_value}{exclude blocks with a combined (replicate 1 +
replicate 2) read count or normalized read count of less than
 \code{combined_min_value} (default is 20 reads)}

\item{combined_max_value}{exclude blocks with a combined (replicate 1 +
replicate 2) read count or normalized read count of more than
 \code{combined_max_value} (default is infinity)}

\item{max_factor}{numeric; controls the replacement values for \code{Inf}
and \code{-Inf}. \code{Inf} are replaced by \code{max(x) * max_factor} and
\code{-Inf} are replaced by \code{min(x) / max_factor}.}

\item{jitter_factor}{numeric; controls the magnitude of the noise that
is added to \code{x}. This is done to break ties in \code{x}.}

\item{mu}{ a starting value for the mean of the reproducible component.}

\item{sigma}{ a starting value for the standard deviation of the
    reproducible component.}

\item{rho}{ a starting value for the correlation coefficient of the
    reproducible component.}

\item{p}{ a starting value for the proportion of reproducible component.}

\item{eps}{ Stopping criterion. Iterations stop when the increment of
    log-likelihood is < eps*log-likelihood, Default=0.001.}

\item{max_iteration}{integer; maximum number of iterations for
IDR estimation (defaults to 30)}

\item{local_idr}{see \code{\link[idr:est.IDR]{est.IDR}}}

\item{use_python}{if Python is not on PATH, specify path to Python binary
here (see \code{\link[reticulate:use_python]{use_python}})}

\item{use_virtualenv}{if Python package \code{hic-straw} is not in base
virtualenv environment, specify environment here (see
\code{\link[reticulate:use_python]{use_virtualenv}})}

\item{use_condaenv}{if Python package \code{hic-straw} is not in base
conda environment, specify environment here (see
\code{\link[reticulate:use_python]{use_condaenv}})}
}
\value{
Data frames \code{rep1_df} and \code{rep2_df} with
the following columns:
\tabular{rll}{
  column 1: \tab \code{chr_a} \tab character; genomic location of anchor A -
  chromosome (e.g., \code{"chr3"})\cr
  column 2: \tab \code{start_a} \tab integer; genomic location of anchor A -
  start coordinate\cr
  column 3: \tab \code{end_a} \tab integer; genomic location of anchor A -
  end coordinate\cr
  column 4: \tab \code{chr_b} \tab character; genomic location of anchor B -
  chromosome (e.g., \code{"chr3"})\cr
  column 5: \tab \code{start_b} \tab integer; genomic location of anchor B -
  start coordinate\cr
  column 6: \tab \code{end_b} \tab integer; genomic location of anchor B -
  end coordinate\cr
  column 7: \tab \code{value} \tab numeric; p-value, FDR, or heuristic used
  to rank the interactions\cr
  column 8: \tab \code{"rep_value"} \tab numeric; value of corresponding
  replicate interaction. If no corresponding interaction was found,
  \code{rep_value} is set to \code{NA}.\cr
  column 9: \tab \code{"rank"} \tab integer; rank of the interaction,
  established by value column, ascending order\cr
  column 10: \tab \code{"rep_rank"} \tab integer; rank of corresponding
  replicate interaction. If no corresponding interaction was found,
  \code{rep_rank} is set to \code{NA}.\cr
  column 11: \tab \code{"idx"} \tab integer; interaction index,
  primary key\cr
  column 12: \tab \code{"rep_idx"} \tab integer; specifies the index of the
  corresponding interaction in the other replicate (foreign key). If no
  corresponding interaction was found, \code{rep_idx} is set to \code{NA}.
}
}
\description{
This method estimates Irreproducible Discovery Rates (IDR) of
genomic interactions between two replicates of HiC experiments.

\code{estimate_idr2d_hic} uses the Python package \code{hic-straw} internally
 to read .hic contact matrix files (see
\href{https://pypi.org/project/hic-straw/}{hic-straw on PyPI} or the
\href{https://pypi.org/project/hic-straw/}{Aiden lab GitHub repository}
for more information).

The contact matrix is subdivided into blocks, where the block size is
determined by \code{resolution}. The reads per block are used to rank blocks
and replicate blocks are easily matched by genomic location.
}
\references{
Q. Li, J. B. Brown, H. Huang and P. J. Bickel. (2011) Measuring
reproducibility of high-throughput experiments. Annals of Applied
Statistics, Vol. 5, No. 3, 1752-1779.
}
