% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{estimateIDR1D}
\alias{estimateIDR1D}
\title{Estimates IDR for Genomic Peak Data}
\usage{
estimateIDR1D(rep1.df, rep2.df, value.transformation = c("identity",
  "additive.inverse", "multiplicative.inverse", "log",
  "log.additive.inverse"), ambiguity.resolution.method = c("overlap",
  "midpoint", "value"), remove.nonstandard.chromosomes = TRUE,
  max.factor = 1.5, jitter.factor = 1e-04, max.gap = 0L, mu = 0.1,
  sigma = 1, rho = 0.2, p = 0.5, eps = 0.001, max.iteration = 30,
  local.idr = TRUE)
}
\arguments{
\item{value.transformation}{the values in \code{x} have to be transformed in
a way such that when ordered in descending order, more significant
interactions end up on top of the list. If the values in \code{x} are
p-values, \code{"log.additive.inverse"} is recommended. The following
transformations are supported:
\tabular{rl}{
  \code{"identity"} \tab no transformation is performed on \code{x}\cr
  \code{"additive.inverse"} \tab \code{x = -x}\cr
  \code{"multiplicative.inverse"} \tab \code{x = 1 / x}\cr
  \code{"log"} \tab \code{x = log(x)}. Note: zeros are replaced by
  \code{.Machine$double.xmin}\cr
  \code{"log.additive.inverse"} \tab \code{x = -log(x)}, recommended if
  \code{x} are p-values. Note: zeros are replaced by
  \code{.Machine$double.xmin}
}

either \code{"ascending"} (more significant
interactions have lower value in \code{value} column) or \code{"descending"}
(more significant interactions have higher value in \code{value} column)}

\item{remove.nonstandard.chromosomes}{removes peaks and interactions
containing
genomic locations on non-standard chromosomes using
\code{\link[GenomeInfoDb:seqlevels-wrappers]{keepStandardChromosomes}}
(default is TRUE)}

\item{max.factor}{numeric; controls the replacement values for \code{Inf}
and \code{-Inf}. \code{Inf} are replaced by \code{max(x) * max.factor} and
\code{-Inf} are replaced by \code{min(x) / max.factor}.}

\item{jitter.factor}{numeric; controls the magnitude of the noise that
is added to \code{x}. This is done to break ties in \code{x}.}

\item{mu}{ a starting value for the mean of the reproducible component.}

\item{sigma}{ a starting value for the standard deviation of the
    reproducible component.}

\item{rho}{ a starting value for the correlation coefficient of the
    reproducible component.}

\item{p}{ a starting value for the proportion of reproducible component.}

\item{eps}{ Stopping criterion. Iterations stop when the increment of
    log-likelihood is < eps*log-likelihood, Default=0.001.}

\item{max.iteration}{integer; maximum number of iterations for
IDR estimation (defaults to 30)}

\item{local.idr}{TODO}
}
\value{
List with two components (\code{rep1.df} and \code{rep1.df})
containing the peaks from input data frames \code{rep1.df} and
\code{rep2.df} with
the following columns:
\tabular{rll}{
  column 1: \tab \code{chr} \tab character; genomic location of peak -
  chromosome (e.g., \code{"chr3"})\cr
  column 2: \tab \code{start} \tab integer; genomic location of peak  -
  start coordinate\cr
  column 3: \tab \code{end} \tab integer; genomic location of peak -
  end coordinate\cr
  column 4: \tab \code{value} \tab numeric; p-value, FDR, or heuristic used
  to rank the peaks\cr
  column 5: \tab \code{rep.value} \tab numeric; value of corresponding
  replicate peak If no corresponding peak was found, \code{rep.value} is set
  to \code{NA}.\cr
  column 6: \tab \code{rank} \tab integer; rank of the peak, established by
  value column, ascending order\cr
  column 7: \tab \code{rep.rank} \tab integer; rank of corresponding
  replicate peak. If no corresponding peak was found, \code{rep.rank} is
  set to \code{NA}.\cr
  column 8: \tab \code{idx} \tab integer; peak index, primary key\cr
  column 9: \tab \code{rep.idx} \tab integer; specifies the index of the
  corresponding peak in the other replicate (foreign key). If no
  corresponding peak was found, \code{rep.idx} is set to \code{NA}.\cr
  column 10: \tab \code{idr} \tab IDR of the peak and the
  corresponding peak in the other replicate. If no corresponding
  peak was found, \code{idr} is set to \code{NA}.
}
}
\description{
Estimates IDR for Genomic Peak Data
}
\examples{
# TODO

}
\references{
Q. Li, J. B. Brown, H. Huang and P. J. Bickel. (2011) Measuring
reproducibility of high-throughput experiments. Annals of Applied
Statistics, Vol. 5, No. 3, 1752-1779.
}
