% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{IDRDistributionHistogram}
\alias{IDRDistributionHistogram}
\title{Create histogram of IDR values}
\usage{
IDRDistributionHistogram(df, remove.na = TRUE, xlab = "IDR",
  ylab = "density", title = "IDR value distribution")
}
\arguments{
\item{df}{part of output of \code{estimateIDR}, a data frame with at
least the following named columns:
\tabular{rl}{
  \code{idr} \tab IDR of the peak and the
  corresponding peak in the other replicate.
}}

\item{remove.na}{logical; should NA values be removed?}

\item{xlab}{character; x axis label}

\item{ylab}{character; y axis label}

\item{title}{character; plot title}
}
\value{
ggplot2 object; IDR distribution histogram
}
\description{
Creates diagnostic plots to visualize the results of \code{estimateIDR}.
}
\examples{
idr.results <- estimateIDR1D(idr2d:::chipseq$rep1.df,
                             idr2d:::chipseq$rep2.df,
                             value.transformation = "log")
IDRDistributionHistogram(idr.results$rep1.df)

}
