% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{rankIDRScatterplot}
\alias{rankIDRScatterplot}
\title{Create scatterplot of IDR values}
\usage{
rankIDRScatterplot(df, remove.na = TRUE, xlab = "rank in replicate 1",
  ylab = "rank in replicate 2", title = "rank - IDR dependence",
  color.gradient = c("rainbow", "default"), max.points.shown = 2500)
}
\arguments{
\item{df}{part of output of \code{estimateIDR}, a data frame with at
least the following named columns:
\tabular{rl}{
  \code{rank} \tab integer; rank of the peak, established by
  value column, ascending order\cr
  \code{rep.rank} \tab integer; rank of corresponding
  replicate peak.\cr
  \code{idr} \tab IDR of the peak and the
  corresponding peak in the other replicate.
}}

\item{remove.na}{logical; should NA values be removed?}

\item{xlab}{character; x axis label}

\item{ylab}{character; y axis label}

\item{title}{character; plot title}

\item{color.gradient}{character; either "rainbow" or "default"}

\item{max.points.shown}{integer; default is 2500}
}
\value{
ggplot2 object; IDR rank scatterplot
}
\description{
Creates diagnostic plots to visualize the results of \code{estimateIDR}.
}
\examples{
idr.results <- estimateIDR1D(idr2d:::chipseq$rep1.df,
                             idr2d:::chipseq$rep2.df,
                             value.transformation = "log")
rankIDRScatterplot(idr.results$rep1.df)

}
