% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{removeNonstandardChromosomes1D}
\alias{removeNonstandardChromosomes1D}
\title{Removes Peaks on Non-standard Chromosomes}
\usage{
removeNonstandardChromosomes1D(x)
}
\arguments{
\item{x}{data frame of genomic peaks, with the following columns
(position of columns matter, column names are irrelevant):
\tabular{rl}{
  column 1 (\code{chr}) \tab character; genomic location of peak -
  chromosome (e.g., \code{"chr3"})\cr
  column 2 \code{start}) \tab integer; genomic location of peak  -
  start coordinate\cr
  column 3 (\code{end}) \tab integer; genomic location of peak -
  end coordinate\cr
  column 4 (\code{value}) \tab numeric; p-value, FDR, or heuristic used to
  rank the peaks
}}
}
\description{
Removes Peaks on Non-standard Chromosomes
}
