% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{removeNonstandardChromosomes2D}
\alias{removeNonstandardChromosomes2D}
\title{Removes Interactions on Non-standard Chromosomes}
\usage{
removeNonstandardChromosomes2D(x)
}
\arguments{
\item{x}{data frame of genomic interactions, with the following columns
(position of columns matter, column names are irrelevant):
\tabular{rl}{
  column 1 (\code{chr.a}) \tab character; genomic location of anchor A -
  chromosome (e.g., \code{"chr3"})\cr
  column 2 \code{start.a}) \tab integer; genomic location of anchor A -
  start coordinate\cr
  column 3 (\code{end.a}) \tab integer; genomic location of anchor A -
  end coordinate\cr
  column 4 (\code{chr.b}) \tab character; genomic location of anchor B -
  chromosome (e.g., \code{"chr3"})\cr
  column 5 (\code{start.b}) \tab integer; genomic location of anchor B -
  start coordinate\cr
  column 6 (\code{end.c}) \tab integer; genomic location of anchor B -
  end coordinate\cr
  column 7 (\code{value}) \tab numeric; p-value, FDR, or heuristic used to
  rank the interactions
}}
}
\description{
Removes Interactions on Non-standard Chromosomes
}
