% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chiapet}
\alias{chiapet}
\title{Example Genomic Interaction Data Set}
\format{
A list with two components, the data frames \code{rep1_df} and
\code{rep2_df}, which have the following seven columns:
\tabular{rll}{
  column 1: \tab \code{chr_a} \tab character; genomic location of anchor A -
  chromosome (e.g., \code{"chr3"})\cr
  column 2: \tab \code{start_a} \tab integer; genomic location of anchor A -
  start coordinate\cr
  column 3: \tab \code{end_a} \tab integer; genomic location of anchor A -
  end coordinate\cr
  column 4: \tab \code{chr_b} \tab character; genomic location of anchor B -
  chromosome (e.g., \code{"chr3"})\cr
  column 5: \tab \code{start_b} \tab integer; genomic location of anchor B -
  start coordinate\cr
  column 6: \tab \code{end_b} \tab integer; genomic location of anchor B -
  end coordinate\cr
  column 7: \tab \code{fdr} \tab numeric; False Discovery Rate -
  significance of interaction
}
}
\usage{
chiapet
}
\description{
This object contains genomic interactions on chromosomes 1 to 5, which could
be the results of Hi-C or ChIA-PET experiments, done in duplicates.
}
\keyword{datasets}
