% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{establishBijection}
\alias{establishBijection}
\title{Finds One-to-One Correspondence between Peaks or interactions
from Replicate 1 and 2}
\usage{
establishBijection(rep1.df, rep2.df, analysis.type = c("IDR1D", "IDR2D"),
  ambiguity.resolution.method = c("overlap", "midpoint", "value"),
  max.gap = 0L)
}
\arguments{
\item{analysis.type}{"IDR2D" for genomic interaction data sets,
"IDR1D" for genomic peak data sets}
}
\description{
Finds One-to-One Correspondence between Peaks or interactions
from Replicate 1 and 2
}
