% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hic.R
\name{estimate_idr2d_hic}
\alias{estimate_idr2d_hic}
\title{Estimates IDR for Genomic Interactions measured by HiC experiments}
\usage{
estimate_idr2d_hic(rep1_df, rep2_df, combined_min_value = 30,
  combined_max_value = Inf, max_factor = 1.5, jitter_factor = 1e-04,
  mu = 0.1, sigma = 1, rho = 0.2, p = 0.5, eps = 0.001,
  max_iteration = 30, local_idr = TRUE)
}
\arguments{
\item{rep1_df}{data frame of parsed .hic file for replicate 1 (output of
\code{\link{parse_hic_file}})}

\item{rep2_df}{data frame of parsed .hic file for replicate 2 (output of
\code{\link{parse_hic_file}})}

\item{combined_min_value}{exclude blocks with a combined (replicate 1 +
replicate 2) read count or normalized read count of less than
 \code{combined_min_value} (default is 20 reads)}

\item{combined_max_value}{exclude blocks with a combined (replicate 1 +
replicate 2) read count or normalized read count of more than
 \code{combined_max_value} (default is infinity)}

\item{max_factor}{numeric; controls the replacement values for \code{Inf}
and \code{-Inf}. \code{Inf} are replaced by \code{max(x) * max_factor} and
\code{-Inf} are replaced by \code{min(x) / max_factor}.}

\item{jitter_factor}{numeric; controls the magnitude of the noise that
is added to \code{x}. This is done to break ties in \code{x}.}

\item{mu}{ a starting value for the mean of the reproducible component.}

\item{sigma}{ a starting value for the standard deviation of the
    reproducible component.}

\item{rho}{ a starting value for the correlation coefficient of the
    reproducible component.}

\item{p}{ a starting value for the proportion of reproducible component.}

\item{eps}{ Stopping criterion. Iterations stop when the increment of
    log-likelihood is < eps*log-likelihood, Default=0.001.}

\item{max_iteration}{integer; maximum number of iterations for
IDR estimation (defaults to 30)}

\item{local_idr}{see \code{\link[idr:est.IDR]{est.IDR}}}
}
\value{
Data frame with the following columns:
\tabular{rll}{
  column 1: \tab \code{interaction} \tab character; genomic location of
  interaction block (e.g., \code{"chr1:204940000-204940000"})\cr
  column 2: \tab \code{value} \tab numeric; p-value, FDR, or heuristic used
  to rank the interactions\cr
  column 3: \tab \code{"rep_value"} \tab numeric; value of corresponding
  replicate interaction\cr
  column 4: \tab \code{"rank"} \tab integer; rank of the interaction,
  established by value column, ascending order\cr
  column 5: \tab \code{"rep_rank"} \tab integer; rank of corresponding
  replicate interaction\cr
  column 6: \tab \code{"idr"} \tab integer; IDR of the block and the
  corresponding block in the other replicate
}
}
\description{
This method estimates Irreproducible Discovery Rates (IDR) of
genomic interactions between two replicates of HiC experiments.

The contact matrix is subdivided into blocks, where the block size is
determined by \code{resolution}. The reads per block are used to rank blocks
and replicate blocks are easily matched by genomic location.
}
\examples{
idr_results <- estimate_idr2d_hic(idr2d:::hic$rep1_df,
                                  idr2d:::hic$rep2_df)
summary(idr_results)

}
\references{
Q. Li, J. B. Brown, H. Huang and P. J. Bickel. (2011) Measuring
reproducibility of high-throughput experiments. Annals of Applied
Statistics, Vol. 5, No. 3, 1752-1779.
}
