% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{estimateIDR}
\alias{estimateIDR}
\title{Estimates IDR for Genomic Interaction Data}
\usage{
estimateIDR(rep1.df, rep2.df, value.transformation = c("identity",
  "additive.inverse", "multiplicative.inverse", "log",
  "log.additive.inverse"), ambiguity.resolution.method = c("value",
  "overlap", "midpoint"), max.factor = 1.5, jitter.factor = 1e-04,
  max.gap = 1000L, mu = 0.1, sigma = 1, rho = 0.2, p = 0.5,
  eps = 0.001, max.iteration = 30, local.idr = TRUE)
}
\arguments{
\item{rep1.df}{data frame of observations (i.e., genomic interactions) of
replicate 1, with at least the following columns (position of columns
matter, column names are irrelevant):
\tabular{rl}{
  column 1 (\code{chrA}) \tab character; genomic location of anchor A -
  chromosome (e.g., \code{"chr3"})\cr
  column 2 \code{startA}) \tab integer; genomic location of anchor A -
  start coordinate\cr
  column 3 (\code{endA}) \tab integer; genomic location of anchor A -
  end coordinate\cr
  column 4 (\code{chrB}) \tab character; genomic location of anchor B -
  chromosome (e.g., \code{"chr3"})\cr
  column 5 (\code{startB}) \tab integer; genomic location of anchor B -
  start coordinate\cr
  column 6 (\code{endB}) \tab integer; genomic location of anchor B -
  end coordinate\cr
  column 7 (\code{value}) \tab numeric; p-value, FDR, or heuristic used to
  rank the interactions
}}

\item{rep2.df}{data frame of observations (i.e., genomic interactions) of
replicate 2, with the following columns (position of columns
matter, column names are irrelevant):
\tabular{rl}{
  column 1 (\code{chrA}) \tab character; genomic location of anchor A -
  chromosome (e.g., \code{"chr3"})\cr
  column 2 \code{startA}) \tab integer; genomic location of anchor A -
  start coordinate\cr
  column 3 (\code{endA}) \tab integer; genomic location of anchor A -
  end coordinate\cr
  column 4 (\code{chrB}) \tab character; genomic location of anchor B -
  chromosome (e.g., \code{"chr3"})\cr
  column 5 (\code{startB}) \tab integer; genomic location of anchor B -
  start coordinate\cr
  column 6 (\code{endB}) \tab integer; genomic location of anchor B -
  end coordinate\cr
  column 7 (\code{value}) \tab numeric; p-value, FDR, or heuristic used to
  rank the interactions
}}

\item{value.transformation}{the values in \code{x} have to be transformed in
a way such that when ordered in descending order, more significant
interactions end up on top of the list. If the values in \code{x} are
p-values, \code{"log.additive.inverse"} is recommended. The following
transformations are supported:
\tabular{rl}{
  \code{"identity"} \tab no transformation is performed on \code{x}\cr
  \code{"additive.inverse"} \tab \code{x = -x}\cr
  \code{"multiplicative.inverse"} \tab \code{x = 1 / x}\cr
  \code{"log"} \tab \code{x = log(x)}. Note: zeros are replaced by
  \code{.Machine$double.xmin}\cr
  \code{"log.additive.inverse"} \tab \code{x = -log(x)}, recommended if
  \code{x} are p-values. Note: zeros are replaced by
  \code{.Machine$double.xmin}
}

either \code{"ascending"} (more significant
interactions have lower value in \code{value} column) or \code{"descending"}
(more significant interactions have higher value in \code{value} column)}

\item{ambiguity.resolution.method}{defines how ambiguous assignments
(when one interaction in replicate 1 overlaps with multiple interactions in
replicate 2 or vice versa)
are resolved. Available methods:
\tabular{rl}{
  \code{"value"} \tab interactions are prioritized by ascending or descending
  \code{value} column (see \code{sorting.direction}), e.g., if two
  interactions in replicate 1 overlap with one interaction in replicate 2,
  the interaction from replicate 1 is chosen which has a lower (if
  \code{sorting.direction} is \code{"ascending"}) or higher (if
  \code{"descending"}) value \cr
  \code{"overlap"} \tab the interaction pair is chosen which has the highest
  relative overlap, i.e., overlap in nucleotides of replicate 1 interaction
  anchor A and replicate 2 interaction anchor A,
  plus replicate 1 interaction anchor B and replicate 2 interaction anchor B,
  normalized by their lengths\cr
  \code{"midpoint"} \tab the interaction pair is chosen which has the
  smallest
  distance between their anchor midpoints, i.e., distance from midpoint of
  replicate 1 interaction anchor A to midpoint of
  replicate 2 interaction anchor A, plus distance from midpoint of
  replicate 1 interaction anchor B to midpoint of
  replicate 2 interaction anchor B
}}

\item{max.factor}{numeric; controls the replacement values for \code{Inf}
and \code{-Inf}. \code{Inf} are replaced by \code{max(x) * max.factor} and
\code{-Inf} are replaced by \code{min(x) / max.factor}.}

\item{jitter.factor}{numeric; controls the magnitude of the noise that
is added to \code{x}. This is done to break ties in \code{x}.}

\item{max.gap}{integer; maximum gap in nucleotides allowed between two
anchors for
them to be considered as overlapping (defaults to \code{1000L})}

\item{mu}{ a starting value for the mean of the reproducible component.}

\item{sigma}{ a starting value for the standard deviation of the
    reproducible component.}

\item{rho}{ a starting value for the correlation coefficient of the
    reproducible component.}

\item{p}{ a starting value for the proportion of reproducible component.}

\item{eps}{ Stopping criterion. Iterations stop when the increment of
    log-likelihood is < eps*log-likelihood, Default=0.001.}

\item{max.iteration}{integer; maximum number of iterations for
IDR estimation (defaults to 30)}

\item{local.idr}{TODO}
}
\value{
List with two components (\code{rep1.df} and \code{rep1.df})
containing the original data frames \code{rep1.df} and \code{rep2.df} with
additional column \code{idr}, which holds the IDR of the interaction and the
corresponding interaction in the other replicate. If no corresponding
interaction was found, \code{idr} is set to \code{NA}.
}
\description{
This method estimates Irreproducible Discovery Rates (IDR) between
two replicates of experiments identifying genomic interactions, such as
HiC, ChIA-PET, and HiChIP.
}
\examples{
idr.results <- estimateIDR(idr2d:::chiapet$rep1.df,
                           idr2d:::chiapet$rep2.df,
                           value.transformation = "log.additive.inverse")

}
\references{
Q. Li, J. B. Brown, H. Huang and P. J. Bickel. (2011) Measuring
reproducibility of high-throughput experiments. Annals of Applied
Statistics, Vol. 5, No. 3, 1752-1779.
}
